+++ agent/src/main/java/com/teamscale/jacoco/agent/options/AgentOptionsParser.java
@@ -14,10 +14,11 @@ import com.teamscale.jacoco.agent.commandline.Validator;
 import com.teamscale.jacoco.agent.git_properties.GitPropertiesLocator;
 import com.teamscale.jacoco.agent.git_properties.InvalidGitPropertiesException;
 import com.teamscale.jacoco.agent.sapnwdi.NwdiConfiguration;
+import com.teamscale.jacoco.agent.upload.artifactory.ArtifactoryConfig;
 import com.teamscale.report.EDuplicateClassFileBehavior;
 import com.teamscale.report.util.BashFileSkippingInputStream;
 import com.teamscale.report.util.ILogger;
-
+import okhttp3.HttpUrl;
 import org.conqat.lib.commons.collections.CollectionUtils;
 import org.conqat.lib.commons.filesystem.FileSystemUtils;
 
@@ -32,8 +33,6 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.jar.JarInputStream;
 import java.util.jar.Manifest;
-
-import okhttp3.HttpUrl;
 
 /**
  * Parses agent command line options.
@@ -106,6 +105,9 @@ public class AgentOptionsParser {
 		} else if (key.startsWith("teamscale-") && handleTeamscaleOptions(options, key, value)) {
 			return;
 		} else if (handleHttpServerOptions(options, key, value)) {
+			return;
+		} else if (key.startsWith("artifactory-") && ArtifactoryConfig
+				.handleArtifactoryOptions(options.artifactoryConfig, filePatternResolver, key, value)) {
 			return;
 		} else if (key.startsWith("azure-") && handleAzureFileStorageOptions(options, key, value)) {
 			return;
@@ -243,7 +245,7 @@ public class AgentOptionsParser {
 						filePatternResolver.parsePath(key, value).toFile());
 				return true;
 			case AgentOptions.TEAMSCALE_GIT_PROPERTIES_JAR_OPTION:
-				options.teamscaleServer.revision = parseGitProperties(key, value);
+				options.teamscaleServer.revision = getRevisionFromGitProperties(key, value);
 				return true;
 			case "teamscale-message":
 				options.teamscaleServer.message = value;
@@ -261,14 +263,14 @@ public class AgentOptionsParser {
 		}
 	}
 
-	private String parseGitProperties(String key, String value) throws AgentOptionParseException {
-		File jarFile = filePatternResolver.parsePath(key, value).toFile();
+	private String getRevisionFromGitProperties(String optionName, String value) throws AgentOptionParseException {
+		File jarFile = filePatternResolver.parsePath(optionName, value).toFile();
 		try {
-			String commit = GitPropertiesLocator.getCommitFromGitProperties(jarFile);
-			if (commit == null) {
+			String revision = GitPropertiesLocator.getRevisionFromGitProperties(jarFile);
+			if (revision == null) {
 				throw new AgentOptionParseException("Could not locate a git.properties file in " + jarFile.toString());
 			}
-			return commit;
+			return revision;
 		} catch (IOException | InvalidGitPropertiesException e) {
 			throw new AgentOptionParseException("Could not locate a valid git.properties file in " + jarFile.toString(),
 					e);