+++ report-generator/src/main/java/com/teamscale/report/jacoco/CoverageFile.java
@@ -46,6 +46,11 @@ public class CoverageFile {
 		return FileSystemUtils.getFilenameWithoutExtension(coverageFile);
 	}
 
+	/** Get the filename of the coverage file. */
+	public String getName() {
+		return coverageFile.getName();
+	}
+
 	/**
 	 * Delete the coverage file from disk
 	 */